# swagger-spec.json
{"openapi":"3.0.0","paths":{"/running":{"get":{"operationId":"AppController_getVersionEndpoint","summary":"Version running","parameters":[],"responses":{"200":{"description":"Endpoint executed with sucess"}}}},"/GoldenRaspberryAwards":{"get":{"operationId":"GRAController_getAll","summary":"Busca filmes","parameters":[],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpExceptionDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpExceptionDto"}}}}},"tags":["Golden Raspberry Awards"]},"post":{"operationId":"GRAController_postRanking","summary":"Grava filmes","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GRAFilmesDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GRAFilmesDTO"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpExceptionDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpExceptionDto"}}}}},"tags":["Golden Raspberry Awards"]}},"/GoldenRaspberryAwards/producer-intervals":{"get":{"operationId":"GRAController_getProducerIntervals","summary":"Lista Produção que Ganharam o premio","parameters":[],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpExceptionDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpExceptionDto"}}}}},"tags":["Golden Raspberry Awards"]}}},"info":{"title":"archref-golden-raspberry-awards","description":"Pior Filme do Golden Raspberry Awards.","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"HttpExceptionDto":{"type":"object","properties":{"statusCode":{"type":"number","description":"Status Code","example":500},"timestamp":{"type":"string","description":"Timestamp","example":"2024-04-08T22:09:20.721Z"},"path":{"type":"string","description":"Requested Path","example":"/health"}},"required":["statusCode","timestamp","path"]},"GRAFilmesDTO":{"type":"object","properties":{"id":{"type":"number","description":"Código indentificador"},"year":{"type":"string","description":"Ano de Lançamento"},"title":{"type":"string","description":"Título do Filme"},"studios":{"type":"string","description":"Estudio"},"producers":{"type":"string","description":"Produtores"},"winner":{"type":"string","description":"Vencedor"}},"required":["id","year","title","studios","producers","winner"]}}}}